+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Release 0.4.0
+
+- [x] Added URLSessionConfiguration builder
+- [x] Other builders refactor
+
 # Release 0.3.6
 
 - [x] Added earliestBeginDate delay to NetURLSession background tasks